split common shield composables
